Update primary key data type in the `kategori` table to VARCHAR(5)
This commit updates the data type of the primary key `id` in the `kategori` table from INT to VARCHAR(5). This change was made to better suit the requirements of the application and improve data storage efficiency. The primary key remains unique and is limited to a maximum length of 5 characters. All related constraints and foreign keys have been adjusted accordingly to maintain data integrity.